refactor: move extract_roi_voxel_tacs_from_image_using_mask to new module
