+++ petpal/preproc/__init__.py
@@ -1,3 +1,4 @@
+"""Tools for preparing PET data for kinetic modeling and visualization"""
 from . import image_operations_4d
 from . import motion_corr
 from . import partial_volume_corrections
@@ -5,7 +6,7 @@ from . import register
 from . import symmetric_geometric_transfer_matrix
 from . import segmentation_tools
 from . import decay_correction
-
+from . import regional_tac_extraction
 
 def main():
     print("PETPAL - Pre-processing")
