+++ petpal/preproc/image_operations_4d.py
@@ -31,7 +31,7 @@ import nibabel
 import numpy as np
 from scipy.ndimage import center_of_mass
 
-from ..utils.useful_functions import weighted_series_sum, check_physical_space_for_ants_image_pair
+from ..utils.useful_functions import weighted_series_sum
 from ..utils import image_io, math_lib
 from ..preproc.decay_correction import undo_decay_correction, decay_correct
 
@@ -620,39 +620,6 @@ def write_tacs(input_image_path: str,
         else:
             out_tac_path = os.path.join(out_tac_dir, f'seg-{regions_abrev[i]}_tac.tsv')
         np.savetxt(out_tac_path,region_tac_file,delimiter='\t',header=header_text,comments='')
-
-
-def extract_roi_voxel_tacs_from_image_using_mask(input_image: ants.core.ANTsImage,
-                                                 mask_image: ants.core.ANTsImage,
-                                                 verbose: bool = False) -> np.ndarray:
-    """
-    Function to extract ROI voxel tacs from an image using a mask image.
-
-    This function returns all the voxel TACs, and unlike :func:`extract_mean_roi_tac_from_nifti_using_segmentation`,
-    does not calculate the mean over all the voxels.
-
-    Args:
-        input_image (ants.core.ANTsImage): Input 4D-image from which to extract ROI voxel tacs.
-        mask_image (ants.core.ANTsImage): Mask image which determines which voxels to extract.
-        verbose (bool, optional): If True, prints information about the shape of extracted voxel tacs.
-
-    Returns:
-        out_voxels (np.ndarray): Array of voxel TACs of shape (num_voxels, num_frames)
-
-    Raises:
-         AssertionError: If input image is not 4D-image.
-         AssertionError: If mask image is not in the same physical space as the input image.
-
-    """
-    assert len(input_image.shape) == 4, "Input image must be 4D."
-    assert check_physical_space_for_ants_image_pair(input_image, mask_image), (
-        "Images must have the same physical dimensions.")
-
-    x_inds, y_inds, z_inds = mask_image.nonzero()
-    out_voxels = input_image.numpy()[x_inds, y_inds, z_inds, :]
-    if verbose:
-        print(f"(ImageOps): Output TACs have shape {out_voxels.shape}")
-    return out_voxels
 
 
 class SimpleAutoImageCropper(object):
